Move gutter paint methods out of lapce-data
