+++ lapce-data/src/editor.rs
@@ -45,7 +45,7 @@ use druid::{
     PaintCtx, Point, Rect, RenderContext, Size, Target, Vec2, WidgetId,
 };
 use druid::{Application, ExtEventSink, MouseEvent};
-use lapce_core::syntax::Syntax;
+pub use lapce_core::syntax::Syntax;
 use lapce_rpc::buffer::BufferId;
 use lsp_types::CompletionTextEdit;
 use lsp_types::{
@@ -1758,511 +1758,6 @@ impl LapceEditorBufferData {
             self.main_split.diagnostics.get_mut(path).map(Arc::make_mut)
         } else {
             None
-        }
-    }
-
-    fn paint_gutter_inline_diff(
-        &self,
-        ctx: &mut PaintCtx,
-        compare: &str,
-        gutter_width: f64,
-    ) {
-        if self.buffer.history_changes.get(compare).is_none() {
-            return;
-        }
-        let self_size = ctx.size();
-        let rect = self_size.to_rect();
-        let changes = self.buffer.history_changes.get(compare).unwrap();
-        let line_height = self.config.editor.line_height as f64;
-        let scroll_offset = self.editor.scroll_offset;
-        let start_line = (scroll_offset.y / line_height).floor() as usize;
-        let end_line =
-            (scroll_offset.y + rect.height() / line_height).ceil() as usize;
-        let current_line = self.editor.cursor.current_line(&self.buffer);
-        let last_line = self.buffer.last_line();
-        let width = self.config.editor_text_width(ctx.text(), "W");
-
-        let mut line = 0;
-        for change in changes.iter() {
-            match change {
-                DiffLines::Left(r) => {
-                    let len = r.len();
-                    line += len;
-
-                    if line < start_line {
-                        continue;
-                    }
-                    ctx.fill(
-                        Size::new(self_size.width, line_height * len as f64)
-                            .to_rect()
-                            .with_origin(Point::new(
-                                0.0,
-                                line_height * (line - len) as f64 - scroll_offset.y,
-                            )),
-                        self.config
-                            .get_color_unchecked(LapceTheme::SOURCE_CONTROL_REMOVED),
-                    );
-                    for l in line - len..line {
-                        if l < start_line {
-                            continue;
-                        }
-                        let actual_line = l - (line - len) + r.start;
-
-                        let content = actual_line + 1;
-                        let x = ((last_line + 1).to_string().len()
-                            - content.to_string().len())
-                            as f64
-                            * width;
-                        let y = line_height * l as f64 + 5.0 - scroll_offset.y;
-                        let pos = Point::new(x, y);
-
-                        let text_layout = ctx
-                            .text()
-                            .new_text_layout(
-                                content.to_string()
-                                    + &vec![
-                                        " ";
-                                        (last_line + 1).to_string().len() + 2
-                                    ]
-                                    .join("")
-                                    + " -",
-                            )
-                            .font(
-                                self.config.editor.font_family(),
-                                self.config.editor.font_size as f64,
-                            )
-                            .text_color(
-                                self.config
-                                    .get_color_unchecked(LapceTheme::EDITOR_DIM)
-                                    .clone(),
-                            )
-                            .build()
-                            .unwrap();
-                        ctx.draw_text(&text_layout, pos);
-
-                        if l > end_line {
-                            break;
-                        }
-                    }
-                }
-                DiffLines::Both(left, r) => {
-                    let len = r.len();
-                    line += len;
-                    if line < start_line {
-                        continue;
-                    }
-
-                    for l in line - len..line {
-                        if l < start_line {
-                            continue;
-                        }
-                        let left_actual_line = l - (line - len) + left.start;
-                        let right_actual_line = l - (line - len) + r.start;
-
-                        let left_content = left_actual_line + 1;
-                        let x = ((last_line + 1).to_string().len()
-                            - left_content.to_string().len())
-                            as f64
-                            * width;
-                        let y = line_height * l as f64 + 5.0 - scroll_offset.y;
-                        let pos = Point::new(x, y);
-
-                        let text_layout = ctx
-                            .text()
-                            .new_text_layout(left_content.to_string())
-                            .font(
-                                self.config.editor.font_family(),
-                                self.config.editor.font_size as f64,
-                            )
-                            .text_color(
-                                self.config
-                                    .get_color_unchecked(LapceTheme::EDITOR_DIM)
-                                    .clone(),
-                            )
-                            .build()
-                            .unwrap();
-                        ctx.draw_text(&text_layout, pos);
-
-                        let right_content = right_actual_line + 1;
-                        let x = ((last_line + 1).to_string().len()
-                            - right_content.to_string().len())
-                            as f64
-                            * width
-                            + gutter_width
-                            + 2.0 * width;
-                        let pos = Point::new(x, y);
-                        let text_layout = ctx
-                            .text()
-                            .new_text_layout(right_content.to_string())
-                            .font(
-                                self.config.editor.font_family(),
-                                self.config.editor.font_size as f64,
-                            )
-                            .text_color(if right_actual_line == current_line {
-                                self.config
-                                    .get_color_unchecked(
-                                        LapceTheme::EDITOR_FOREGROUND,
-                                    )
-                                    .clone()
-                            } else {
-                                self.config
-                                    .get_color_unchecked(LapceTheme::EDITOR_DIM)
-                                    .clone()
-                            })
-                            .build()
-                            .unwrap();
-                        ctx.draw_text(&text_layout, pos);
-
-                        if l > end_line {
-                            break;
-                        }
-                    }
-                }
-                DiffLines::Skip(_l, _r) => {
-                    let rect = Size::new(self_size.width, line_height)
-                        .to_rect()
-                        .with_origin(Point::new(
-                            0.0,
-                            line_height * line as f64 - scroll_offset.y,
-                        ));
-                    ctx.fill(
-                        rect,
-                        self.config
-                            .get_color_unchecked(LapceTheme::PANEL_BACKGROUND),
-                    );
-                    ctx.stroke(
-                        rect,
-                        self.config
-                            .get_color_unchecked(LapceTheme::EDITOR_FOREGROUND),
-                        1.0,
-                    );
-                    line += 1;
-                }
-                DiffLines::Right(r) => {
-                    let len = r.len();
-                    line += len;
-                    if line < start_line {
-                        continue;
-                    }
-
-                    ctx.fill(
-                        Size::new(self_size.width, line_height * len as f64)
-                            .to_rect()
-                            .with_origin(Point::new(
-                                0.0,
-                                line_height * (line - len) as f64 - scroll_offset.y,
-                            )),
-                        self.config
-                            .get_color_unchecked(LapceTheme::SOURCE_CONTROL_ADDED),
-                    );
-
-                    for l in line - len..line {
-                        if l < start_line {
-                            continue;
-                        }
-                        let actual_line = l - (line - len) + r.start;
-
-                        let content = actual_line + 1;
-                        let x = ((last_line + 1).to_string().len()
-                            - content.to_string().len())
-                            as f64
-                            * width
-                            + gutter_width
-                            + 2.0 * width;
-                        let y = line_height * l as f64 + 5.0 - scroll_offset.y;
-                        let pos = Point::new(x, y);
-
-                        let text_layout = ctx
-                            .text()
-                            .new_text_layout(content.to_string() + " +")
-                            .font(
-                                self.config.editor.font_family(),
-                                self.config.editor.font_size as f64,
-                            )
-                            .text_color(if actual_line == current_line {
-                                self.config
-                                    .get_color_unchecked(
-                                        LapceTheme::EDITOR_FOREGROUND,
-                                    )
-                                    .clone()
-                            } else {
-                                self.config
-                                    .get_color_unchecked(LapceTheme::EDITOR_DIM)
-                                    .clone()
-                            })
-                            .build()
-                            .unwrap();
-                        ctx.draw_text(&text_layout, pos);
-
-                        if l > end_line {
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    fn paint_gutter_code_lens(&self, ctx: &mut PaintCtx, _gutter_width: f64) {
-        let rect = ctx.size().to_rect();
-        let scroll_offset = self.editor.scroll_offset;
-        let empty_lens = Syntax::lens_from_normal_lines(
-            self.buffer.len(),
-            self.config.editor.line_height,
-            self.config.editor.code_lens_font_size,
-            &[],
-        );
-        let lens = if let Some(syntax) = self.buffer.syntax.as_ref() {
-            &syntax.lens
-        } else {
-            &empty_lens
-        };
-
-        let cursor_line = self
-            .buffer
-            .line_of_offset(self.editor.cursor.offset().min(self.buffer.len()));
-        let last_line = self.buffer.line_of_offset(self.buffer.len());
-        let start_line = lens
-            .line_of_height(scroll_offset.y.floor() as usize)
-            .min(last_line);
-        let end_line = lens
-            .line_of_height(
-                (scroll_offset.y + rect.height()).ceil() as usize
-                    + self.config.editor.line_height,
-            )
-            .min(last_line);
-        let char_width = self
-            .config
-            .char_width(ctx.text(), self.config.editor.font_size as f64);
-        let max_line_width = (last_line + 1).to_string().len() as f64 * char_width;
-
-        let mut y = lens.height_of_line(start_line) as f64;
-        for (line, line_height) in lens.iter_chunks(start_line..end_line + 1) {
-            let content = if *self.main_split.active != Some(self.view_id)
-                || self.editor.cursor.is_insert()
-                || line == cursor_line
-            {
-                line + 1
-            } else if line > cursor_line {
-                line - cursor_line
-            } else {
-                cursor_line - line
-            };
-            let content = content.to_string();
-            let is_small = line_height < self.config.editor.line_height;
-            let text_layout = ctx
-                .text()
-                .new_text_layout(content.clone())
-                .font(
-                    self.config.editor.font_family(),
-                    if is_small {
-                        self.config.editor.code_lens_font_size as f64
-                    } else {
-                        self.config.editor.font_size as f64
-                    },
-                )
-                .text_color(if line == cursor_line {
-                    self.config
-                        .get_color_unchecked(LapceTheme::EDITOR_FOREGROUND)
-                        .clone()
-                } else {
-                    self.config
-                        .get_color_unchecked(LapceTheme::EDITOR_DIM)
-                        .clone()
-                })
-                .build()
-                .unwrap();
-            let x = max_line_width - text_layout.size().width;
-            let pos = Point::new(
-                x,
-                y - scroll_offset.y
-                    + if is_small {
-                        0.0
-                    } else {
-                        (line_height as f64 - text_layout.size().height) / 2.0
-                    },
-            );
-            ctx.draw_text(&text_layout, pos);
-
-            y += line_height as f64;
-        }
-    }
-
-    pub fn paint_gutter(&self, ctx: &mut PaintCtx, gutter_width: f64) {
-        let rect = ctx.size().to_rect();
-        ctx.with_save(|ctx| {
-            let clip_rect = rect;
-            ctx.clip(clip_rect);
-            if let Some(compare) = self.editor.compare.as_ref() {
-                self.paint_gutter_inline_diff(ctx, compare, gutter_width);
-                return;
-            }
-            if self.editor.code_lens {
-                self.paint_gutter_code_lens(ctx, gutter_width);
-                return;
-            }
-            let line_height = self.config.editor.line_height as f64;
-            let scroll_offset = self.editor.scroll_offset;
-            let start_line = (scroll_offset.y / line_height).floor() as usize;
-            let end_line =
-                (scroll_offset.y + rect.height() / line_height).ceil() as usize;
-            let num_lines = (ctx.size().height / line_height).floor() as usize;
-            let last_line = self.buffer.last_line();
-            let current_line = self.editor.cursor.current_line(&self.buffer);
-            let width = self.config.editor_text_width(ctx.text(), "W");
-            for line in start_line..start_line + num_lines + 1 {
-                if line > last_line {
-                    break;
-                }
-                let content = if *self.main_split.active != Some(self.view_id)
-                    || self.editor.cursor.is_insert()
-                    || line == current_line
-                {
-                    line + 1
-                } else if line > current_line {
-                    line - current_line
-                } else {
-                    current_line - line
-                };
-                let content = content.to_string();
-
-                let text_layout = ctx
-                    .text()
-                    .new_text_layout(content.clone())
-                    .font(
-                        self.config.editor.font_family(),
-                        self.config.editor.font_size as f64,
-                    )
-                    .text_color(if line == current_line {
-                        self.config
-                            .get_color_unchecked(LapceTheme::EDITOR_FOREGROUND)
-                            .clone()
-                    } else {
-                        self.config
-                            .get_color_unchecked(LapceTheme::EDITOR_DIM)
-                            .clone()
-                    })
-                    .build()
-                    .unwrap();
-                let x = ((last_line + 1).to_string().len() - content.len()) as f64
-                    * width;
-                let y = line_height * line as f64 - scroll_offset.y
-                    + (line_height - text_layout.size().height) / 2.0;
-                let pos = Point::new(x, y);
-                ctx.draw_text(&text_layout, pos);
-            }
-
-            if let Some(changes) = self.buffer.history_changes.get("head") {
-                let mut line = 0;
-                let mut last_change = None;
-                for change in changes.iter() {
-                    let len = match change {
-                        DiffLines::Left(_range) => 0,
-                        DiffLines::Skip(_left, right) => right.len(),
-                        DiffLines::Both(_left, right) => right.len(),
-                        DiffLines::Right(range) => range.len(),
-                    };
-                    line += len;
-                    if line < start_line {
-                        last_change = Some(change.clone());
-                        continue;
-                    }
-
-                    let mut modified = false;
-                    let color = match change {
-                        DiffLines::Left(_range) => {
-                            Some(self.config.get_color_unchecked(
-                                LapceTheme::SOURCE_CONTROL_REMOVED,
-                            ))
-                        }
-                        DiffLines::Right(_range) => {
-                            if let Some(DiffLines::Left(_)) = last_change.as_ref() {
-                                modified = true;
-                            }
-                            if modified {
-                                Some(self.config.get_color_unchecked(
-                                    LapceTheme::SOURCE_CONTROL_MODIFIED,
-                                ))
-                            } else {
-                                Some(self.config.get_color_unchecked(
-                                    LapceTheme::SOURCE_CONTROL_ADDED,
-                                ))
-                            }
-                        }
-                        _ => None,
-                    };
-
-                    if let Some(color) = color {
-                        let removed_height = 10.0;
-                        let size = Size::new(
-                            3.0,
-                            if len == 0 {
-                                removed_height
-                            } else {
-                                line_height * len as f64
-                            },
-                        );
-                        let x = gutter_width + width;
-                        let mut y =
-                            (line - len) as f64 * line_height - scroll_offset.y;
-                        if len == 0 {
-                            y -= removed_height / 2.0;
-                        }
-                        if modified {
-                            let rect = Size::new(3.0, removed_height)
-                                .to_rect()
-                                .with_origin(Point::new(
-                                    x,
-                                    y - removed_height / 2.0,
-                                ));
-                            ctx.fill(
-                                rect,
-                                self.config.get_color_unchecked(
-                                    LapceTheme::EDITOR_BACKGROUND,
-                                ),
-                            );
-                        }
-                        let rect = size.to_rect().with_origin(Point::new(x, y));
-                        ctx.fill(rect, &color.clone().with_alpha(0.8));
-                    }
-
-                    if line > end_line {
-                        break;
-                    }
-                    last_change = Some(change.clone());
-                }
-            }
-
-            if *self.main_split.active == Some(self.view_id) {
-                self.paint_code_actions_hint(ctx, gutter_width);
-            }
-        });
-    }
-
-    fn paint_code_actions_hint(&self, ctx: &mut PaintCtx, gutter_width: f64) {
-        if let Some(actions) = self.current_code_actions() {
-            if !actions.is_empty() {
-                let line_height = self.config.editor.line_height as f64;
-                let offset = self.editor.cursor.offset();
-                let (line, _) = self
-                    .buffer
-                    .offset_to_line_col(offset, self.config.editor.tab_width);
-                let svg = get_svg("lightbulb.svg").unwrap();
-                let width = 16.0;
-                let height = 16.0;
-                let char_width = self.config.editor_text_width(ctx.text(), "W");
-                let rect =
-                    Size::new(width, height).to_rect().with_origin(Point::new(
-                        gutter_width + char_width + 3.0,
-                        (line_height - height) / 2.0 + line_height * line as f64
-                            - self.editor.scroll_offset.y,
-                    ));
-                ctx.draw_svg(
-                    &svg,
-                    rect,
-                    Some(self.config.get_color_unchecked(LapceTheme::LAPCE_WARN)),
-                );
-            }
         }
     }
